Improved the install script in two ways: - made apt-get update optional (if you have to debug it really slows down the processÃ) - made installing bitcoin optional, and now it doesn't break if ~/.bitcoin already exists

+++ install.sh
@@ -13,27 +13,41 @@ do
     shift
 done
 
-sudo apt-get update
+read -p "Do you want to update? [y/N]? " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    sudo apt-get update
+fi
 sudo apt-get install python-dev vim screen
 sudo pip install -r requirements.txt
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 HOME="$DIR/.."
 
-wget --directory-prefix=$HOME https://bitcoin.org/bin/0.9.1/bitcoin-0.9.1-linux.tar.gz
-tar -C $HOME -zxvf $HOME/bitcoin-0.9.1-linux.tar.gz
-mv $HOME/bitcoin-0.9.1-linux $HOME/bitcoin
-rm $HOME/bitcoin-0.9.1-linux.tar.gz
 
-cp $DIR/src/settings_local.py.example $DIR/src/settings_local.py
+read -p "Do you need to install bitcoind? [y/N]? " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    wget --directory-prefix=$HOME https://bitcoin.org/bin/0.9.1/bitcoin-0.9.1-linux.tar.gz &&
+    tar -C $HOME -zxvf $HOME/bitcoin-0.9.1-linux.tar.gz &&
+    mv $HOME/bitcoin-0.9.1-linux $HOME/bitcoin &&
+    rm $HOME/bitcoin-0.9.1-linux.tar.gz &&
+    echo 'alias bitcoind=~/bitcoin/bin/64/bitcoin' >> $HOME/.bash_aliases &&
+    source $HOME/.bash_aliases &&
+    
+    cp $DIR/src/settings_local.py.example $DIR/src/settings_local.py
+fi
+
 
 if [ "$tflag" == "yes" ]
 then
   echo BITCOIND_TEST_MODE=True >> $DIR/src/settings_local.py
 fi
 
-
-mkdir $HOME/.bitcoin/
+mkdir -p $HOME/.bitcoin/
+# this is harmless even if the file exists
 touch $HOME/.bitcoin/bitcoin.conf
 
 BTCRPC=`openssl rand -hex 32`
